chore Closed #20: extract components from new-ticket page
